tuan7/requirements.txt
connexion[swagger-ui]==2.14.2
flask-cors==4.0.1
pymongo==4.8.0
dnspython==2.8.0
python-dotenv==1.0.1
